backend/requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
python-dotenv
fastapi
uvicorn